created cards folder, renamed files, moved to new folder
