+++ app/src/components/dashboard_control.js
