+++ app/src/components/leave_group.js
